+++ tests/integration_tests/saas/test_stripe_task.py
@@ -1071,9 +1071,8 @@ def test_stripe_erasure_request_task(
         ],
     )
 
-    # run erasure with MASKING_STRICT to execute the update actions
-
-    config.execution.MASKING_STRICT = True
+    # Run erasure with masking_strict = False so both update and delete actions can be used
+    config.execution.MASKING_STRICT = False
 
     x = graph_task.run_erasure(
         privacy_request,
@@ -1084,20 +1083,20 @@ def test_stripe_erasure_request_task(
         get_cached_data_for_erasures(privacy_request.id),
     )
 
-    # verify masking request was issued for endpoints with an update action
+    # verify masking request was issued for endpoints with both update/delete actions
     assert x == {
         f"{dataset_name}:customer": 1,
-        f"{dataset_name}:tax_id": 0,
-        f"{dataset_name}:invoice_item": 0,
+        f"{dataset_name}:tax_id": 1,
+        f"{dataset_name}:invoice_item": 1,
         f"{dataset_name}:charge": 0,
-        f"{dataset_name}:invoice": 0,
+        f"{dataset_name}:invoice": 2,
         f"{dataset_name}:card": 1,
         f"{dataset_name}:customer_balance_transaction": 0,
         f"{dataset_name}:payment_intent": 0,
         f"{dataset_name}:payment_method": 2,
         f"{dataset_name}:credit_note": 0,
         f"{dataset_name}:bank_account": 1,
-        f"{dataset_name}:subscription": 0,
+        f"{dataset_name}:subscription": 1,
         f"{dataset_name}:dispute": 0,
     }
 
@@ -1146,46 +1145,6 @@ def test_stripe_erasure_request_task(
     bank_account = response.json()["data"][0]
     assert bank_account["account_holder_name"] == "MASKED"
 
-    # run erasure without MASKING_STRICT to execute the delete actions
-
-    config.execution.MASKING_STRICT = False
-
-    x = graph_task.run_erasure(
-        privacy_request,
-        erasure_policy_string_rewrite,
-        graph,
-        [stripe_connection_config],
-        {"email": stripe_erasure_identity_email},
-        get_cached_data_for_erasures(privacy_request.id),
-    )
-
-    # verify masking request was issued for endpoints with an update or delete action
-    assert x == {
-        f"{dataset_name}:customer": 1,
-        f"{dataset_name}:tax_id": 1,
-        f"{dataset_name}:invoice_item": 1,
-        f"{dataset_name}:charge": 0,
-        f"{dataset_name}:invoice": 2,
-        f"{dataset_name}:card": 1,
-        f"{dataset_name}:customer_balance_transaction": 0,
-        f"{dataset_name}:payment_intent": 0,
-        f"{dataset_name}:payment_method": 2,
-        f"{dataset_name}:credit_note": 0,
-        f"{dataset_name}:bank_account": 1,
-        f"{dataset_name}:subscription": 1,
-        f"{dataset_name}:dispute": 0,
-    }
-
-    # customer
-    response = requests.get(
-        url=f"{base_url}/v1/customers",
-        headers=headers,
-        params={"email": stripe_erasure_identity_email},
-    )
-    customer = response.json()["data"][0]
-    customer_id = customer["id"]
-    assert customer["shipping"]["name"] == "MASKED"
-
     # tax_id
     response = requests.get(
         url=f"{base_url}/v1/customers/{customer_id}/tax_ids", headers=headers
@@ -1202,34 +1161,6 @@ def test_stripe_erasure_request_task(
     invoice_item = response.json()["data"]
     # Can't delete an invoice item that is attached to an invoice that is no longer editable
     assert len(invoice_item) == 1
-
-    # card
-    response = requests.get(
-        url=f"{base_url}/v1/customers/{customer_id}/sources",
-        headers=headers,
-        params={"object": "card"},
-    )
-    card = response.json()["data"][0]
-    assert card["name"] == "MASKED"
-
-    # payment_method
-    response = requests.get(
-        url=f"{base_url}/v1/customers/{customer_id}/payment_methods",
-        headers=headers,
-        params={"type": "card"},
-    )
-    payment_methods = response.json()["data"]
-    for payment_method in payment_methods:
-        assert payment_method["billing_details"]["name"] == "MASKED"
-
-    # bank_account
-    response = requests.get(
-        url=f"{base_url}/v1/customers/{customer_id}/sources",
-        headers=headers,
-        params={"object": "bank_account"},
-    )
-    bank_account = response.json()["data"][0]
-    assert bank_account["account_holder_name"] == "MASKED"
 
     # subscription
     response = requests.get(